% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin.R
\name{isValidLipidName}
\alias{isValidLipidName}
\title{Check the provided lipid name against the built-in grammars.
Will return FALSE if none of the parsers was able to parse the provided name successfully.}
\usage{
isValidLipidName(lipidName)
}
\arguments{
\item{lipidName}{The lipid name to check.}
}
\value{
TRUE if the lipidName could be parsed, FALSE otherwise.
}
\description{
Check the provided lipid name against the built-in grammars.
Will return FALSE if none of the parsers was able to parse the provided name successfully.
}
\examples{
isValidLipidName("PC 32:1")
isValidLipidName("PC(32:1)")
isValidLipidName("PCX(32:1)")
}
