% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin.R
\name{parseLipidName}
\alias{parseLipidName}
\title{Parse the provided lipid name and return structural information.
Will return an empty vector if none of the parsers was able to parse the provided name successfully.}
\usage{
parseLipidName(lipidName)
}
\arguments{
\item{lipidName}{the lipid name to check.}
}
\description{
Parse the provided lipid name and return structural information.
Will return an empty vector if none of the parsers was able to parse the provided name successfully.
}
