% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin-package.R
\docType{package}
\name{rgoslin-package}
\alias{rgoslin}
\alias{rgoslin-package}
\title{rgoslin: Lipid name parsing and normalization}
\description{
rGoslin is the R implementation for the Grammar of Succint Lipid Nomenclature.
}
\details{
rgoslin is the R implementation of the "grammar of succint lipid nomenclature". It provides 
implementations for parsing of lipid names from LipidMAPS, SwissLipids and other common sources of lipid names.
}
\references{
NA
}
\author{
\strong{Maintainer}: Nils Hoffmann \email{nils.hoffmann@isas.de}

Authors:
\itemize{
  \item Dominik Kopczynski \email{dominik.kopczynski@isas.de}
}

}
\keyword{internal}
