% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin.R
\name{isValidLipidName}
\alias{isValidLipidName}
\title{Check the provided lipid name against the built-in grammars.
Will return FALSE if none of the parsers was able to parse the provided name successfully.}
\usage{
isValidLipidName(lipidName)
}
\arguments{
\item{lipidName}{the lipid name to check.}
}
\description{
Check the provided lipid name against the built-in grammars.
Will return FALSE if none of the parsers was able to parse the provided name successfully.
}
