% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin-package.R
\docType{package}
\name{rgoslin-package}
\alias{rgoslin}
\alias{rgoslin-package}
\title{rgoslin: Lipid Name Parsing and Normalization}
\description{
The R implementation for the Grammar of Succint Lipid Nomenclature parses different short hand notation dialects for lipid names and normalizes them to a standard name and provides calculated monoisotopic masses and sum formulas.
}
\details{
rgoslin is the R implementation of the "grammar of succint lipid nomenclature". It provides 
implementations for parsing of shorthand lipid names from LipidMAPS, SwissLipids and other common sources of lipid names.
}
\references{
NA
}
\author{
\strong{Maintainer}: Nils Hoffmann \email{nils.hoffmann@cebitec.uni-bielefeld.de}

Authors:
\itemize{
  \item Dominik Kopczynski \email{dominik.kopczynski@isas.de}
}

}
\keyword{internal}
